post api/GpsExercise get api/GpsExercise/{id} delete api/GpsExercise/{id} done, tested, working. Database updated for cascade deletion in GpsPath in case of deleting from GpsExercise.

+++ testAPI/Models/GpsExercise.cs
@@ -1,19 +1,18 @@
 ﻿namespace FitnessApp.Models
 {
     using System;
-    using System.Collections.Generic;
 
     public class GpsExercise
     {
         public int UserID { get; set; }
         public string Type { get; set; }
         public DateTime Start { get; set; }
-        public float Duration { get; set; }
-        public float AvgSpeed { get; set; }
-        public float Burned { get; set; }
+        public double Duration { get; set; }
+        public double AvgSpeed { get; set; }
+        public double Burned { get; set; }
         public GpsPath[] Path {get;set;}
 
-        public GpsExercise(int userID, string type, DateTime start, float duration, float avgSpeed, float burned, GpsPath[] path)
+        public GpsExercise(int userID, string type, DateTime start, double duration, double avgSpeed, double burned, GpsPath[] path)
         {
             this.UserID = userID;
             this.Type = type;
